+++ test/Base.js
@@ -1,4 +1,3 @@
-import "./Async/TaskTest.js";
 import "./Caching/MemoryCacheTest.js";
 import "./Data/PaginationTest.js";
 import "./Data/SortTest.js";
@@ -8,6 +7,7 @@ import "./Net/IPAddressTest.js";
 import "./Net/MailAddressTest.js";
 import "./Net/PhoneNumberTest.js";
 import "./Security/AccessLevelTest.js";
+import "./Threading/TaskTest.js";
 import "./Util/DateTest.js";
 import "./Util/NumberTest.js";
 import "./Util/StringTest.js";