+++ src/UI/PageSelector.ts
@@ -1,4 +1,4 @@
-import {Pagination} from "#Data/Pagination.js";
+import {PaginationState} from "#Web/PaginationState.js";
 import {html, type TemplateResult} from "lit";
 import {customElement, property} from "lit/decorators.js";
 import {classMap} from "lit/directives/class-map.js";
@@ -44,7 +44,7 @@ export class PageSelector extends Component {
 	/**
 	 * The underlying pagination model.
 	 */
-	@property({attribute: false}) pagination = new Pagination;
+	@property({attribute: false}) pagination = new PaginationState;
 
 	/**
 	 * The number of items displayed around the active item.
@@ -55,15 +55,15 @@ export class PageSelector extends Component {
 	 * The list of buttons to display.
 	 */
 	get #items(): number[] {
-		const start = Math.max(1, this.pagination.page - this.sideItems);
-		const end = Math.min(this.pagination.pageCount, this.pagination.page + this.sideItems);
+		const start = Math.max(1, this.pagination.currentPage - this.sideItems);
+		const end = Math.min(this.pagination.lastPage, this.pagination.currentPage + this.sideItems);
 
 		const items = [];
 		if (start > 1) items.push(1);
 		if (start > 2) items.push(0);
 		for (let index = start; index <= end; index++) items.push(index);
-		if (end < this.pagination.pageCount - 1) items.push(0);
-		if (end < this.pagination.pageCount) items.push(this.pagination.pageCount);
+		if (end < this.pagination.lastPage - 1) items.push(0);
+		if (end < this.pagination.lastPage) items.push(this.pagination.lastPage);
 		return items;
 	}
 
@@ -74,7 +74,7 @@ export class PageSelector extends Component {
 	 */
 	navigate(page: number, event?: Event): void {
 		if (event) (event.target as Element).closest("button")?.blur();
-		if (page >= 1 && page <= this.pagination.pageCount && page != this.pagination.page) this.dispatchEvent(new PageSelectorEvent(page));
+		if (page >= 1 && page <= this.pagination.lastPage && page != this.pagination.currentPage) this.dispatchEvent(new PageSelectorEvent(page));
 	}
 
 	/**
@@ -82,23 +82,23 @@ export class PageSelector extends Component {
 	 * @returns The view template.
 	 */
 	protected override render(): TemplateResult {
-		return when(this.pagination.pageCount > 1, () => html`
+		return when(this.pagination.lastPage > 1, () => html`
 			<nav>
 				<ul class="pagination justify-content-${this.alignment}">
-					<li class="page-item ${classMap({disabled: this.pagination.page <= 1})}">
-						<button class="page-link" @click=${(event: Event) => this.navigate(this.pagination.page - 1, event)}>
+					<li class="page-item ${classMap({disabled: this.pagination.currentPage <= 1})}">
+						<button class="page-link" @click=${(event: Event) => this.navigate(this.pagination.currentPage - 1, event)}>
 							<i class="icon">chevron_left</i><span class="d-none d-sm-inline ms-1">Précédent</span>
 						</button>
 					</li>
 					${this.#items.map(index => html`
-						<li class="page-item ${classMap({active: index == this.pagination.page, disabled: !index})}">
+						<li class="page-item ${classMap({active: index == this.pagination.currentPage, disabled: !index})}">
 							<button class="page-link" @click=${(event: Event) => this.navigate(index, event)}>
 								${index ? index : "..."}
 							</button>
 						</li>
 					`)}
-					<li class="page-item ${classMap({disabled: this.pagination.page >= this.pagination.pageCount})}">
-						<button class="page-link" @click=${(event: Event) => this.navigate(this.pagination.page + 1, event)}>
+					<li class="page-item ${classMap({disabled: this.pagination.currentPage >= this.pagination.lastPage})}">
+						<button class="page-link" @click=${(event: Event) => this.navigate(this.pagination.currentPage + 1, event)}>
 							<span class="d-none d-sm-inline me-1">Suivant</span><i class="icon">chevron_right</i>
 						</button>
 					</li>