+++ test/Base.js
@@ -1,5 +1,4 @@
 import "./Caching/MemoryCacheTest.js";
-import "./Data/SortTest.js";
 import "./DependencyInjection/ContainerTest.js";
 import "./Globalization/RegionInfoTest.js";
 import "./Net/IPAddressTest.js";
@@ -11,3 +10,4 @@ import "./Util/DateTest.js";
 import "./Util/NumberTest.js";
 import "./Util/StringTest.js";
 import "./Web/PaginationStateTest.js";
+import "./Web/SortTest.js";